move errors to `error` module
